# packages/shared/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/products:
    get:
      operationId: ProductsController_list
      parameters: []
      responses:
        "200":
          description: ""
      summary: Public catalogue — the 16 audited products (no auth required)
      tags: &a6
        - products
    post:
      operationId: ProductsController_create
      parameters: []
      responses:
        "201":
          description: ""
      security:
        - bearer: []
      summary: Create a product (admin only — catalogue is curated from the site audit)
      tags: *a6
...
  /api/v1/products/{slug}:
    get:
      operationId: ProductsController_getOne
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Product detail by slug
      tags: *a6
  /api/v1/products/{id}:
    patch:
      operationId: ProductsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDto"
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      summary: Update a product (admin catalog manage)
      tags: *a6
...
  /api/v1/notifications:
    get:
      operationId: NotificationsController_list
      parameters: []
      responses:
        "200":
          description: ""
      security: &a13
        - bearer: []
      summary: List my notifications (newest first)
      tags: &a14
        - notifications
  /api/v1/notifications/unread-count:
    get:
      operationId: NotificationsController_unreadCount
      parameters: []
      responses:
        "200":
          description: ""
      security: *a13
      summary: Unread notification count
      tags: *a14
  /api/v1/notifications/{id}/read:
    patch:
      operationId: NotificationsController_markRead
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a13
      summary: Mark one notification as read (must belong to caller)
      tags: *a14
  /api/v1/notifications/device-tokens:
    post:
      operationId: NotificationsController_registerToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTokenDto"
      responses:
        "201":
          description: ""
      security: *a13
      summary: Register a device push token (multi-device supported)
      tags: *a14
    delete:
      operationId: NotificationsController_removeToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTokenDto"
      responses:
        "200":
          description: ""
      security: *a13
      summary: Remove this device push token
      tags: *a14
  /api/v1/notifications/preferences:
    get:
      operationId: NotificationsController_preferences
      parameters: []
      responses:
        "200":
          description: ""
      security: *a13
      summary: My per-event-type notification preferences
      tags: *a14
  /api/v1/notifications/preferences/{eventType}:
    patch:
      operationId: NotificationsController_updatePreference
      parameters:
        - name: eventType
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferenceDto"
      responses:
        "200":
          description: ""
      security: *a13
      summary: Enable/disable one event type
      tags: *a14
...
  /api/v1/orders:
    get:
      operationId: OrdersController_list
      parameters: []
      responses:
        "200":
          description: ""
      security: &a17
        - bearer: []
      summary: List orders (own for buyers; all for staff/admin)
      tags: &a18
        - orders
  /api/v1/orders/{id}:
    get:
      operationId: OrdersController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a17
      summary: Order detail with items and shipments
      tags: *a18
  /api/v1/orders/{id}/status:
    patch:
      operationId: OrdersController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusDto"
      responses:
        "200":
          description: ""
      security: *a17
      summary: Update order status (staff/admin, strict forward flow)
      tags: *a18
...
info:
  title: EGT B2B API
  description: Eagle Goods Trading Co. — sourcing & export backend API (v1)
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateProductDto:
      type: object
      properties:
        name:
          type: string
        shortDescription:
          type: string
        orderVolume:
          type: string
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: number
...
    RegisterTokenDto:
      type: object
      properties:
        token:
          type: string
        platform:
          type: string
          enum:
            - android
            - ios
            - web
      required:
        - token
        - platform
    UpdatePreferenceDto:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
...
    UpdateOrderStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - confirmed
            - in_production
            - ready_to_ship
            - shipped
            - delivered
            - cancelled
      required:
        - status
...